backend/requirements.txt
Django==3.2.4
django-filter==23.1
django-admin-autocomplete-filter==0.7.1
django-debug-toolbar==3.8.1
djangorestframework==3.12.4
djoser==2.1.0
django-extra-fields==0.3
FastAPI==0.112.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.9.0
PyYAML==6.0
gunicorn==20.1.0